ai_rag_bot/requirements.txt
faiss-cpu
sentence-transformers
openai
pypdf2
